Move multiboot package under the hal package
