+++ kernel/kmain.go
@@ -3,7 +3,7 @@ package kernel
 import (
 	_ "unsafe" // required for go:linkname
 
-	"github.com/achilleasa/gopher-os/kernel/multiboot"
+	"github.com/achilleasa/gopher-os/kernel/hal/multiboot"
 )
 
 // Kmain is the only Go symbol that is visible (exported) from the rt0 initialization
